Move connection decorator to PyaoscxModule
The external and independent function that originally lived in a
separate file was moved to PyaoscxModule to aling the implementation
with the object-oriented approach.

Also, create a new read-only property to read the status of the
connection, based upon the cookies. It is no longer needed to use an
intermediary object attribute to keep the connection state.

Signed-off-by: Javier Albornoz <javier.albornoz@hpe.com>
Change-Id: Ic56bfa3e7005e8c52700263679679478947bed0e