Fix ACE modification requests
There are currently two issues in ACE modifications.

The first one is that the simple comparison between two dictionaries,
each containing an ACE's attributes, to determine a change is
inadequate, because if one of the dictionaries contains null values,
this is detected as a change, when in reality, for PYAOSCX, this means
that the value must be kept as is. An extra check is done now to ensure
only non-null mutable parameters are used to determine if an ACE has
changed.

The second issue is that immutable parameters are getting sent to the
switch in PUT requests when updating ACEs, this is not supported, so
now, if all immutable parameters are unchanged, they are removed from
the data structure that gets sent in the PUT request, if however, there
is at least one immutable attribute changed, the current ACE is deleted
from the switch, and a new one is created with the new value for the
attribute

Change-Id: I8ff135e0e6689fb0e1d9ed78d6a82da6168ac73b