Work around invalid GET response for non-existing Interfaces
Performing a GET request for a non-existent VLAN Interface doesn't
result in a 404 error, instead, a response with

{"other_config": null}

is returned. This causes the factory to not send the POST request
to create the interface because not getting the 404 error is
interpreted as the interface already existing.

The factory has been modified to try a POST before the GET, the
POST fails in case there already is an interface, and then does
the GET to obtain the interface configuration.

CR 201466

Change-Id: I71de106b8014fa8aba72372c85d8e0435e4d4b18