Removed the correct attributes when updating an ACL entry
ACL entries pull mutable and immutable attributes from the switch
when GET requests are executed. When updating the entry only
mutable attributes can be sent as part of the request.
There is a logic error that caused ACL entries to send immutable
attributes as part of update requests, this was fixed such that
only mutable attributes are included in those requests.

CR 199695

Change-Id: Ia5a352322b7cca2a0c439c23843a26e3003afcf5