Fix VSX config_attrs getting reduced
When updating only some VSX attributes, but not all of them, the
config_attrs attribute of the Vsx class gets "reduced" to only the ones
present in the keyword arguments received by PyaoscxFactory.vsx(),
this causes the VSX attributes in the switch to get deleted, because the
PUT request doesn't contain them.

Letting Vsx.get() set the config_attrs list from the GET response fixes
the issue.

CR 202175

Change-Id: I93cbfab259b1e84fa90693ea4613d42dea4a7854