Fix VRF in devices with different capabilities
The original issue presented while attempting to configure an L3
interface. When performing a GET request of an interface on a device
that is not BGP-capable, the VRF module fails because it does not take
into account the device capabilities.

The same issue occurs for OSPF (both v3 and v2). Once this was changed
to take device capabilities into account, a second issue showed up on
some platforms, where an interface using the default VRF returns the
index, and name in the URI of the VRF as an empty string respectively,
that is:

    {"": "/rest/v10.04/system/vrfs"}

instead of:

    {"default": "/rest/v10.04/system/vrfs/default"}

which causes a subsequent GET request to fail.

The changes to fix these issues include:
- Add checks for whether the device is BGP-capable or OSPF-capable in
  Vrf.get().
- Add check for an empty string for the VRF index in Interface.get(),
  and use "default" if that's the case.

Other changes:
- Rearrange imports in device.py to comply with PEP-8
- Remove blank line at beginning of interface.py
- Add check in configure_l3() for the vrf parameter, it previously
  accepted only Vrf objects, despite its docstring saying it could be a
  Vrf name (a string)
- Turn the PyaoscxFactory and Device classes into singletons, via a
  metaclass
- Remove previous code to make PyaoscxFactory a singleton
- Add check to only call device.get() if device.materialized is False
- Remove unused BgpRouter import from vrf.py

Change-Id: I898d5e3e9faf5f34be620f576ff237c84e621e11