Fix config_attrs when calling get()
When get() is called without providing a depth or selector, their
default value is None, this causes the config_attrs list attribute to
not get set after a check on selector fails, this may cause a subsequent
update to not pass all persent values in the object, resulting in
configuration loss.

This is fixed here by using an auxiliary function that performs all
operations needed with depth and selector, including setting the
config_attrs list attribute after the data is pulled from the switch.

The OSPF Router module is also updated to test the change works for
several modules.

CR 205644

Change-Id: I360db720beca5c17d7f968928dbb76a8ec98192c