Fix isl_port Interface instantiation
A VSX's isl_port can be set as an interface name, after creating and
materializing the Vsx object, this can cause an error because the data
sent to the switch can be in the wrong format, if not using an Interface
instance, like was initially intended.

This change makes it so the isl_port attribute is always checked when
sending data to the switch, and transformed into the correct format for
network requests' payloads when the isl_port is present and not null.

Change-Id: I3c0ccb45cfcc5f4d48275eecebe3902c5f748b68