Shorten payloads for network requests
json.dumps() is used throughout the PYAOSCX codebase for serializing
dictionary data into JSON strings for the network request payloads, the
optional sort_keys and indent arguments are currently used for
formatting the strings, this however, increases the size of the request
payloads, and is only really useful for producing human-readable output,
so it's not needed.

This commit removes the use of those optional parameters to shorten the
size of the network requests' payloads, and leaves the instances of
json.dumps() that are not made for formatting a network request's
payload.

Change-Id: I55dd5443ddbe2d1c603924267e59c8bc8d5b774a