Fix config_attrs if calling get() without selector
The original issue was made apparent when reviewing public issue #11 at:
https://github.com/aruba/pyaoscx/issues/11

From there it can be seen that, when get() is called without providing a
selector, its default value is None, which causes the config_attrs list
attribute to not get set after a check on selector fails, this can cause
a subsequent update to not pass all present values in the object,
because that process is dependent on the config_attrs being set, thus
resulting in configuration loss.

This CL fixes this by setting a valid default value in the places where
it wasn't getting set.

Change-Id: I65a6e2ce5bffdb58f2162bb7a213a2a34c1e3ce7