Fix GenericOperationError Exception
The GenericOperationError class is inconsistent in the number of
arguments it requires, it works with either 0, 2, or 3 arguments, but
not only 1, if passed only 1, it raises an Exception.

This CL fixes it so it accepts 0, 1, 2, or 3 arguments, removes an extra
quote character from its dosctring, simplifies its docstring
description, and simplifies how its message is constructed, after
validating its arguments.

Change-Id: Id75575bde8e31b778e50db4dc59cc3571cab9d4a