Fix VerificationError Exception
The VerificationError class is inconsistent in the number of arguments
it takes, it works with either 0, or 2, arguments, if passed only 1
argument, it raises an Exception.

This CL fixes it so it accepts 0 or more arguments, simplifies its
docstring description, and simplifies how its message is constructed,
after validating its arguments.

Change-Id: I14e4bc747466250572acb179ecea1b4afa154d7b