Fix MAC Addresses validation
The sticky MAC addresses' VLANs are not validated.

This CL changes the verification method of the MAC addresses
and adds an error message when an input is invalid. This makes it easier
for the user to identify which MAC address is not valid. Additionally,
the VLANs of the sticky MACs are validated, this parameter should be a
list of integers.

Change-Id: I0239ebf062da8d86276a81e6889c72ba0c280438