Fix VLAN Interface Active Gateway setting
The REST API expects a list of IP addresses, instead of a single IP
address, this causes the type of the value to be "wrong" when passed
from client code.

This CL converts the value to a list of a single element so the REST API
gets the expected type of value.

CR 217361

Change-Id: If29051518288bb3add5a78f7dc8d5003caa0c7dc