Remove default LAG configuration values
The default LAG configuration in L2 and L3 contains default values that
could silently change more than what is intended.

This CL changes the default LAG configuration in those methods, from
"passive" for the lacp mode, and False for the lacp fallback mode and
mc_lag options, to None, and adds an evaluation to see if an actual
value was passed, to either change it, or leave as-is.

Change-Id: Ib066df834b422bd42be5cae78747da0efff10d4f