Change default selector for readonly MAC module
Initially pointed out in https://github.com/aruba/pyaoscx/issues/18,
the MAC module doesn't show information about the port, this stems from
the default selector used in the module, which attempts to pull writable
information, however the REST API resource does not allow configuration,
so the payload is empty, there also appears to be an issue where the MAC
object's path is misconstructed in some cases.

This CL changes the default selector to 'status' to get the expected
behavior when using the get() method, and fixes the path issue by
forcing the mac address to always be lowercase.

Change-Id: I8baafae02a839e0f6feaec248240a0cdebb379f8