Fix ACL entry funcionality
When an ACL entry is created, updated or deleted, a version number
is associated with the parent ACL. This does not allow to configure
the ACL entry correctly, a warning on the switch recommends
reseting the ACL.
Also, to apply the ACL to an interface, the version number is needed,
but with the new implementation it cannot be taken from the ACL.
This change removes the update of this version number. It is not
necessary to configure an ACL entry.
To apply an ACL to the interfaces, the version number is created
separately from the ACL. Also, funcionality to delete the configuration
when applying an ACL to an interface with 'out' direction is added.

CR 236095

Change-Id: Id6cecf98f7ece03632f1451844195348aac00dd3
(cherry picked from commit 3cf9dfe49b78a174b7b751975677f7290c08fefe)