Fix DNS funcionality
DNS setup and delete functions need session argument, since a
configuration object is created, but it is not given. Also, the
functions used on the configuration module try to search for arguments
but they are not found.
This change adds the session argument to create the configuration
object. Also, before asking for specific parameters on the functions
used on the configuration module, a get is performed.

Change-Id: Icc30570882a4ca3df48826e4d97447afe95e896e
(cherry picked from commit 1628d6b991b7de4646d1b35815d3a12d640a54f2)