IP version check with respect to assigned type
When is assigned an IP address in the source or destination IP, there is
no validation of these match with the version type, so it is possible
assign an ipv4 address with a type ipv6.
These changes allow verifying before apply the change the congruence
between the version type and the version of the source and destination
IP. If both are different an error is raised.

CR 251552

Change-Id: I2a0518a775459632ba807c5cbdf2641cf0e30dad
(cherry picked from commit 2d5a014d25cfc892fc15de28fa9fad17acb92344)