Add funcionality to avoid IP overlapping
When adding a new IP address to an interface, there is no validation if
it is overlapping with another IP address on an interface with the same
VRF. Also, incorrect IPs can be send to the switch.

This change adds funcionality that can be applied to multiple modules to
avoid this incorrect behavior when adding a new IP address. The
overlapping function is applied to some use cases on the interface
module.

Change-Id: I97ea37688939ee375c9940f3dca88f1cb2a9071d
(cherry picked from commit b116b475a2865ad4d773d9b396aa75acaf5e9725)