Fix pyaoscx modules for ACL functionality
ACL modules are not working as expected, there are some errors
when the version of the ACL is set. Also the interface attachment
has errors in the configuration.

This commit fixes the errors and the ACL version is set properly.

Change-Id: Idc91c18921c224f017ffc06ee966d7fe35e041bc
(cherry picked from commit d01f89e290e135e16090f32d51f639079ba4e728)