Fix for delete lag from physical interface
When trying to delete the configuration of the physical interface
with the interface module, the lag to which this interface belongs
is remain. This commit add some lines to delete the lag of the
interface and apply this change for shown it correctly on the
running config.

AOSANSI-268637

Change-Id: I2b28e21489c5210e83182f365f37c53cb78a3c88
(cherry picked from commit a439ff77da33cf686bfa0fb9d0f2e6da17db01ef)