Change selector to get capabilities
The "get" method for device was changed in a previous commit and
capabilities are not being downloaded with the default selector.

For interface, vlan and vrf module the selector for device.get
is changed to "status".

Change-Id: I948d773850c5c261f67c6332de0596b6e54100f7
(cherry picked from commit f54eae9132e0c101fa11e97c0cfd9f255a56ae4f)